style: Change slice interface locations
