+++ frontend/src/store.ts
@@ -1,12 +1,12 @@
 import { configureStore } from "@reduxjs/toolkit";
 import authSlice from "./slices/authSlice";
 import apiSlice from "./slices/apiSlice";
-import gameDataSlice from "./slices/gameDataSlice";
+import gameDataSlice, { IGameData } from "./slices/gameDataSlice";
 import throttle from "lodash/throttle";
-import { IGameData } from "../../shared/types";
+import { IAuthData } from "../../shared/types";
 
 export interface IGameState {
-  auth: { userInfo };
+  auth: IAuthData;
   gameData: IGameData;
   completedPuzzles?: { solvedPuzzles };
 }
@@ -17,7 +17,6 @@ function loadState() {
     if (serializedState === null) {
       return undefined;
     }
-
     return JSON.parse(serializedState);
   } catch (e) {
     return undefined;
