+++ shared/types.ts
@@ -66,25 +66,12 @@ export enum SetCompletedStatus {
   incomplete = " (incomplete)"
 }
 
+export interface IAuthData {
+  userInfo: any;
+}
+
 export interface IPuzzleData {
   name: string;
-}
-
-export interface IGameData {
-  numBits: number;
-  totalNumBits: number;
-  currencyAmount: number;
-  userEmail: string;
-  ups: IUpgradesData;
-  savedSolvedPuzzles: string[];
-}
-
-export interface IUpgradesData {
-  acquired: IUpgrade[];
-  purchasable: IUpgrade[];
-  unavailable: IUpgrade[];
-  uncategorized: IUpgrade[];
-  // savedSolvedPuzzles: string[];
 }
 
 export interface IRGBTriple {
